requirements.txt

cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses-json==0.5.2
dnspython==2.0.0
et-xmlfile==1.0.1
fire==0.3.1
Flask==1.1.2
gevent==20.6.2
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
greenlet==0.4.16
grequests==0.6.0
gunicorn==20.0.4
httplib2==0.18.1
hypixel==0.7.9
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-enum==1.5.1
matplotlib==3.3.2
mypy-extensions==0.4.3
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
openpyxl==3.0.5
pandas==1.1.2
Pillow==7.2.0
praw==7.1.0
prawcore==1.5.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-youtube==0.6.4
pytube3==9.6.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
stringcase==1.2.0
termcolor==1.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0